Pass thread errors back to main program thread
Instead of trying to handle any errors themselves, the threads now just
terminate in case of an error and pass the error as their result code to
the main programm thread which is waiting for them to terminate.

There the error will then be reported.

Also, in case of any error, the main program thread will now (try to)
cancel all still-active threads, thus getting them terminated as quickly
as possible.