Simplify error processing
Since error messages will be output as soon as they have been set and
then will be cleared, there is no further reason to check whether an
error message has already been set before setting another one.

This was only necessary when the <error> variable served a twofold
purpose: To store the next error message and as an indication whether or
not any error has occurred at all (to be checked at the end of the
program).