More or less revert to earlier error processing
It turned out that the earlier twofold purpose of the <error> variable
still works, and that clearing the error variable after displaying the
message is not necessary either.

This is because the error message will immediately be displayed after it
has been set, so it does not matter whether it gets overwritten by a new
message later or not.

But even though things are back pretty much as they have been before,
the new ERROR() macro will still be left in place because it makes the
progam shorter.