C99-ify source code
With higher warning levels, it turned out that <inttypes.h> makes use of
C99 features which are not available for plain ANSI C 89. (At least on
GNU platforms, that is.)

Therefore migrate the project to C99, because it needs a C99 compiler
anyway because of <inttypes.h>.

This also gives the opportunity to make use of the %z specifier in
printf() for size_t values, rather than casting to "unsigned long" and
hoping for the best.