Use key file for seeding
The former practice of specifying the seed value directly as an argument
has proven to be problematic for two reasons:

* The seed value is actually a simple (not overly secure) cryptographic
key and needs to be kept for later verification. If the value is only
present in the command line, it may be get lost. By forcing the user to
store it in a file, the chances of losing the seed value are much
smaller.

* Cryptographic keys, no matter how little secure, should never be
passed directly on the command line, because other users on the same
machine might see the command line arguments when running "ps".

Another benefit is the fact that the seed can now be binary and is no
longer restricted to represent text.

The only disadvantage of this change is that existing text files for
storing a seed value will need to be edited in order to remove the
newline character at the end of the (only) line containing a textual
seed value.

This is because all bytes of the seed file will be used for seeding the
PRNG, including (which is normally not intended) any newline character
at its end.