Implement basic buffer comparison code
pearnd_xor() works similar to pearnd_write(), except that the generated
PRNG date is XORed instead of written to the buffer.

Also, pearnd_xor() returns whether or not all XOR operations resulted in
zero values. The latter means that some of the original data bits did
not match the PRNG data.