Add script for creating a block device for testing
A RAM-backed block device will be created which actually contains only a
single sector or actual data. However, before that data sector, a number
of virtual zero-byte sectors will be present, the count of which will be
parsed from the name (or symlink) name of the script.

This allows very large virtual devices in the exabyte range to be
created for testing the starting offset parameter of the mediatester
application.

Replace the explicitly implemented parser for the offset parameter with
an appropriate invocation of sscanf(). This should reduce code size, at
least when linking dynamically to the C runtime library.

Fix selection of the standard file descriptor used for seeking to the
starting offset.

Use symbolic constants for the standard file descriptors.