Use non-local jumps for most of error handling
Although local errors will still be handled directly if they are
unique, errors which have to cross function boundaries or errors which
might occur multiple times at different places in the program will now
be handled using longjmp().

The massively changed code compiles and runs as long as the longjmp() is
not triggered, but the longjmp() handler does not work correctly yet.