Test more thoroughly whether repositioning failed
I have learned that just because lseek() does not give an error, this
does not necessarily mean it worked.

For instance, when seeking /dev/full, this does not give an error. Yet
when reading the current file position back, it always says 0.

Detect this case and abort with an error message. The user should not
specify an explicit offset in such a case.

Also, in case of a write error, display the offset where the error
occurred as well as the number of bytes written successfully before
then.