Check reported I/O block size for validity
The ioctls for determining the sector size of the underlying block
device did not always return meaningful values - in particular, on some
platforms 0 was reported as the sector size.

This led to invalid values for other parameters derived from this, and
so the whole program failed.

Fix this by using the MMU page size or atomic pipe write size as a
fallback if the block device driver reports 0 for whatever reason.