Unconditionally enable Linux-specific features
Define the feature-test macros for Linux before trying to include the
header files.

This program is Linux-specific anyway, so there is no point in letting
the user define those macros manually via -D options.

Also change the "clean" Makefile target so that it deletes the object
files of the helper library also.