Fix type of ioctl result buffers
Debugging has revealed that they seem to be "int" instead of "long".
This made a difference on 64 bit platforms.

I will examine the Kernel source code in order to finally decide on
this, because official documentation for those I/O controls seems hard
to find.

Anyway, changing the types to "int" fixes the problem on all platforms
tested so far, so it is at least an improvement.